- P4A::executeExternalCommands() method was added - all tasks dealing with external systems where moved from P4A::main() to P4A::executeExternalCommands() - session_start() instruction was moved from P4A::singleton() method to p4a.php main inclusion file - P4A::executeExternalCommands() is executed just after the session_start(), in the p4a.php main inclusion file - P4A_Object::destroy() now checks if every object to destroy is a P4A_Object before calling destroy() recursively on it

+++ p4a/p4a_object.php
@@ -156,7 +156,9 @@ abstract class P4A_Object
 		$this_name = $this->getName();
 
 		foreach($this->_objects as $key=>$object_id) {
-			$p4a->objects[$object_id]->destroy();
+			if (isset($p4a->objects[$object_id]) and $p4a->objects[$object_id] instanceof P4A_Object) {
+				$p4a->objects[$object_id]->destroy();
+			}
 			unset($this->_objects[$key]);
 		}
